refactor(benchmark): reuse shared Dot in head-to-head picker
